requirements.txt
google-generativeai
IPython
langchain
langchain_community
langchain_google_genai
pypdf
chromadb